[DEV-98] refactor: 수정
