+++ src/containers/todo/list-todo-component.tsx
@@ -1,11 +1,12 @@
 import FileIcon from "@/assets/file.svg";
 import LinkIcon from "@/assets/link.svg";
 import NoteIcon from "@/assets/note.svg";
-import Goal from "@/containers/todo/Goal";
-import TodoTitleAndCheckBox from "@/containers/todo/TodoTitleAndCheckBox";
+import TodoTitleAndCheckBox from "@/containers/todo/todo-title-checkbox";
 import { DropdownItem, Todo } from "@/types/todo";
 
 import Dropdown from "../../components/dropdown";
+
+import Goal from "./goal";
 
 interface Props {
   todo: Todo;
